+++ examples/Console/Synchronous/EthanI/AI24Bit/AI_on_demand_once.py
@@ -1,0 +1,74 @@
+'''
+AI24Bit - AI_on_demand_once.py with synchronous mode.
+
+This example demonstrates the process of obtaining AI data in on demand mode.
+Additionally, it retrieve AI data from EthanI.
+
+-------------------------------------------------------------------------------------
+Please change correct serial number or IP and port number BEFORE you run example code.
+
+For other examples please check:
+    https://github.com/WPC-Systems-Ltd/WPC_Python_driver_release/tree/main/examples
+See README.md file to get detailed usage of this example.
+
+Copyright (c) 2023 WPC Systems Ltd. All rights reserved.
+'''
+
+## Python
+import time
+
+## WPC
+
+from wpcsys import pywpc
+
+def main():
+    ## Get Python driver version
+    print(f'{pywpc.PKG_FULL_NAME} - Version {pywpc.__version__}')
+
+    ## Create device handle
+    dev = pywpc.EthanI()
+
+    ## Connect to device
+    try:
+        dev.connect("192.168.1.110") ## Depend on your device
+    except Exception as err:
+        pywpc.printGenericError(err)
+        ## Release device handle
+        dev.close()
+        return
+
+    try:
+        ## Parameters setting
+        port = 0 ## Depend on your device
+        timeout = 3  ## second
+
+        ## Get firmware model & version
+        driver_info = dev.Sys_getDriverInfo(timeout=timeout)
+        print("Model name: " + driver_info[0])
+        print("Firmware version: " + driver_info[-1])
+
+        ## Open AI
+        err = dev.AI_open(port, timeout=timeout)
+        print(f"AI_open in port {port}: {err}")
+
+        ## Read data acquisition
+        data =  dev.AI_readOnDemand(port, timeout=timeout)
+        print(f"data in port {port}: ")
+        print(f"{data}")
+
+        ## Close AI
+        err = dev.AI_close(port, timeout=timeout)
+        print(f"AI_close in port {port}: {err}")
+    except Exception as err:
+        pywpc.printGenericError(err)
+
+    ## Disconnect device
+    dev.disconnect()
+
+    ## Release device handle
+    dev.close()
+
+    return
+
+if __name__ == '__main__':
+    main()